jamie first commit 1/5 validators complete
