+++ app.js
@@ -6,6 +6,7 @@ const logger = require('morgan');
 const { sequelize } = require('./db/models');
 const session = require('express-session');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
+
 const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
