+++ authorization.js
@@ -7,16 +7,12 @@ const loginUser = (req, res, user) => {
   };
 };
 
-
-
 const requireAuth = (req, res, next) => {
   if (!res.locals.authenticated) {
     return res.redirect('/user/login');
   }
   return next();
 };
-
-
 
 module.exports = {
   loginUser,
