+++ routes/index.js
@@ -1,9 +1,132 @@
-var express = require('express');
-var router = express.Router();
+const express = require('express');
+const { check, validationResult } = require('express-validator');
+const bcrypt = require('bcryptjs');
 
-/* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render('index', { title: 'a/A Express Skeleton Home' });
+const db = require('../db/models');
+const { csrfProtection, asyncHandler } = require('./utils');
+const { loginUser } = require('../authorization');
+
+const router = express.Router();
+router.get('/', (req, res) => {
+  res.render('index', {title: "QuestTrackr"})
+})
+
+router.get('/register', csrfProtection, (req, res) => {
+  const user = db.User.build();
+  res.render('register', {
+    title: 'Register',
+    user,
+    csrfToken: req.csrfToken(),
+  });
 });
+
+const userValidators = [
+  check('userName')
+    .exists({ checkFalsy: true })
+    .withMessage('Please provide a value for First Name, quest-taker')
+    .isLength({ max: 30, min: 2 })
+    .withMessage('First Name must not be more than 20 characters long'),
+  check('password')
+    .exists({ checkFalsy: true })
+    .withMessage('Please provide a value for Password, quest-taker')
+    .matches(/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])/, 'g')
+    .withMessage('Password must contain at least 1 lowercase letter, uppercase letter and a number, quest-taker'),
+  check('confirmPassword')
+    .exists({ checkFalsy: true })
+    .withMessage('Please provide a value for Confirm Password, quest-taker')
+    .custom((value, { req }) => value === req.body.password)
+    .withMessage('Your password confirmation does not match the entered password, quest-taker')
+];
+
+router.post('/register', csrfProtection, userValidators,
+  asyncHandler(async (req, res) => {
+    const {
+      userName,
+      password,
+      email //this is in database schema
+    } = req.body;
+
+    const user = db.User.build({
+      userName,
+      email //this is in database schema
+    });
+
+    const validatorErrors = validationResult(req);
+
+    if (validatorErrors.isEmpty()) {
+      const hashedPassword = await bcrypt.hash(password, 10);
+      user.hashedPassword = hashedPassword;
+      await user.save();
+      loginUser(req, res, user);
+      res.redirect('/');
+    } else {
+      const errors = validatorErrors.array().map((error) => error.msg);
+      res.render('register', {
+        title: 'Register',
+        user,
+        errors,
+        csrfToken: req.csrfToken(),
+      });
+    }
+  }));
+
+router.get('/login', csrfProtection, (req, res) => {
+  res.render('login', {
+    title: 'Login',
+    csrfToken: req.csrfToken(),
+  });
+});
+
+const loginValidators = [
+  check('userName')
+    .exists({ checkFalsy: true })
+    .withMessage('Please provide a value for a username, quest-taker'),
+  check('password')
+    .exists({ checkFalsy: true })
+    .withMessage('Please provide a value for Password, quest-taker'),
+];
+
+router.post('/login', csrfProtection, loginValidators,
+  asyncHandler(async (req, res) => {
+    const {
+      userName,
+      password,
+    } = req.body;
+
+    let errors = [];
+    const validatorErrors = validationResult(req);
+
+    if (validatorErrors.isEmpty()) {
+      // Attempt to get the user by their user name!
+      const user = await db.User.findOne({ where: { userName } });
+
+      if (user !== null) {
+        // If the user exists then compare their password
+        // to the provided password(quest taker)!
+        const passwordMatch = await bcrypt.compare(password, user.hashedPassword.toString());
+
+        if (passwordMatch) {
+          // If the password hashes match, then login the user name!
+          // and redirect them to the home route.
+          loginUser(req, res, user);
+          return res.redirect('/');
+        }
+      }
+
+      // Otherwise display an error message to the user.
+      errors.push('Login failed for the provided user name and password, quest taker');
+    } else {
+      errors = validatorErrors.array().map((error) => error.msg);
+    }
+
+    res.render('login', {
+      title: 'Login',
+      userName,
+      errors,
+      csrfToken: req.csrfToken(),
+    });
+  }));
+
+
 
 module.exports = router;
